Rename more namespaces
