+++ src/Api/MatchApi.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace HappyrMatch\ApiClient\Api;
 
 use HappyrMatch\ApiClient\Exception;
-use HappyrMatch\ApiClient\Model\Match\CandidateMatch;
-use HappyrMatch\ApiClient\Model\Match\GroupFilter;
-use HappyrMatch\ApiClient\Model\Match\GroupMatch;
+use HappyrMatch\ApiClient\Model\MatchApi\CandidateMatch;
+use HappyrMatch\ApiClient\Model\MatchApi\GroupFilter;
+use HappyrMatch\ApiClient\Model\MatchApi\GroupMatch;
 use Psr\Http\Message\ResponseInterface;
 use Webmozart\Assert\Assert;
