+++ tests/Model/MatchApi/CandidateMatchTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace HappyrMatch\ApiClient\Tests\Model\Role;
+namespace HappyrMatch\ApiClient\Tests\Model\MatchApi;
 
 use HappyrMatch\ApiClient\Model\Match\CandidateMatch;
 use HappyrMatch\ApiClient\Tests\Model\BaseModelTest;
