+++ tests/Model/MatchApi/GroupFilterTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace HappyrMatch\ApiClient\Tests\Model\Role;
+namespace HappyrMatch\ApiClient\Tests\Model\MatchApi;
 
 use HappyrMatch\ApiClient\Model\Group;
 use HappyrMatch\ApiClient\Model\Match\GroupFilter;
